refactor(swing-ui): gather the menu family in its own package
The menu composables sat flat among the widgets, next to Label and ProgressBar,
though they compose into a different tree and are marked for a different applier.
They move to components.menu with a facade of their own, taking the popup that
hosts a menu's composition with them, and their tests move to mirror them.

Nothing about the menus themselves changes: outside the moved files this is imports
and the regenerated API dump.
