+++ docs/COMPONENTS.md
@@ -11,6 +11,7 @@ import androidx.compose.runtime.*
 import org.jetbrains.compose.swing.components.*
 import org.jetbrains.compose.swing.components.button.*
 import org.jetbrains.compose.swing.components.layout.*
+import org.jetbrains.compose.swing.components.menu.*
 import org.jetbrains.compose.swing.components.selection.*
 import org.jetbrains.compose.swing.components.text.*
 import org.jetbrains.compose.swing.modifier.*
@@ -31,6 +32,7 @@ import androidx.compose.runtime.*
 import org.jetbrains.compose.swing.components.*
 import org.jetbrains.compose.swing.components.button.*
 import org.jetbrains.compose.swing.components.layout.*
+import org.jetbrains.compose.swing.components.menu.*
 import org.jetbrains.compose.swing.window.*
 import java.awt.Dimension
