+++ swing-ui/src/test/kotlin/org/jetbrains/compose/swing/modifier/appearance/ButtonAndTextAppearanceModifierTest.kt
@@ -3,11 +3,11 @@ package org.jetbrains.compose.swing.modifier.appearance
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
-import org.jetbrains.compose.swing.components.MenuItem
 import org.jetbrains.compose.swing.components.ProgressBar
 import org.jetbrains.compose.swing.components.button.Button
 import org.jetbrains.compose.swing.components.button.CheckBox
 import org.jetbrains.compose.swing.components.layout.ToolBar
+import org.jetbrains.compose.swing.components.menu.MenuItem
 import org.jetbrains.compose.swing.components.text.TextField
 import org.jetbrains.compose.swing.composeMenu
 import org.jetbrains.compose.swing.modifier.SwingModifier
