+++ swing-ui/src/test/kotlin/org/jetbrains/compose/swing/modifier/interaction/ContextMenuDetachTest.kt
@@ -5,8 +5,8 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import org.jetbrains.compose.swing.components.Label
-import org.jetbrains.compose.swing.components.MenuItem
 import org.jetbrains.compose.swing.components.layout.Column
+import org.jetbrains.compose.swing.components.menu.MenuItem
 import org.jetbrains.compose.swing.menuItemTexts
 import org.jetbrains.compose.swing.modifier.SwingModifier
 import org.jetbrains.compose.swing.modifier.applyModifier
