+++ swing-ui/src/test/kotlin/org/jetbrains/compose/swing/window/WindowMenuBarTest.kt
@@ -3,11 +3,11 @@ package org.jetbrains.compose.swing.window
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
-import org.jetbrains.compose.swing.components.CheckBoxMenuItem
 import org.jetbrains.compose.swing.components.Label
-import org.jetbrains.compose.swing.components.Menu
-import org.jetbrains.compose.swing.components.MenuItem
-import org.jetbrains.compose.swing.components.MenuSeparator
+import org.jetbrains.compose.swing.components.menu.CheckBoxMenuItem
+import org.jetbrains.compose.swing.components.menu.Menu
+import org.jetbrains.compose.swing.components.menu.MenuItem
+import org.jetbrains.compose.swing.components.menu.MenuSeparator
 import org.jetbrains.compose.swing.menuItemTexts
 import org.jetbrains.compose.swing.test.onWindowWithTitle
 import org.jetbrains.compose.swing.test.runComposeSwingTest
